@@ -1,68 +1,45 @@
-module github.com/kuuland/kuu
+module github.com/kuuland/kuu/v3
+
+go 1.21.4
 
 require (
-	github.com/asaskevich/govalidator v0.0.0-20210307081110-f21760c49a8d
-	github.com/boltdb/bolt v1.3.1
 	github.com/buger/jsonparser v1.1.1
-	github.com/dgrijalva/jwt-go v3.2.0+incompatible
-	github.com/ghodss/yaml v1.0.1-0.20220118164431-d8423dcdf344
-	github.com/gin-contrib/cors v1.3.1
-	github.com/gin-contrib/gzip v0.0.5
-	github.com/gin-gonic/gin v1.7.7
-	github.com/go-redis/redis/v8 v8.0.0-beta.7.0.20200728132603-3fbf7df0144b
-	github.com/go-session/gin-session v3.1.0+incompatible
-	github.com/gorilla/websocket v1.5.0
-	github.com/imdario/mergo v0.3.11
-	github.com/jinzhu/gorm v1.9.16
-	github.com/jinzhu/inflection v1.0.0
-	github.com/json-iterator/go v1.1.12
-	github.com/jtolds/gls v4.20.0+incompatible
-	github.com/mojocn/base64Captcha v1.3.1
-	github.com/pkg/errors v0.8.1
-	github.com/robfig/cron/v3 v3.0.0
-	github.com/satori/go.uuid v1.2.0
-	github.com/sirupsen/logrus v1.4.2
-	github.com/stretchr/testify v1.7.0
-	github.com/xuri/excelize/v2 v2.6.0
-	golang.org/x/crypto v0.0.0-20220408190544-5352b0902921
-	gopkg.in/guregu/null.v3 v3.4.0
-	gopkg.in/yaml.v2 v2.3.0
+	github.com/bwmarrin/snowflake v0.3.0
+	github.com/gabriel-vasile/mimetype v1.4.3
+	github.com/go-playground/validator/v10 v10.19.0
+	github.com/gofiber/fiber/v2 v2.52.3
+	github.com/google/uuid v1.6.0
+	github.com/iamdanielyin/strcase v0.0.0-20230814062728-cf61287678ca
+	github.com/joho/godotenv v1.5.1
+	github.com/orandin/lumberjackrus v1.0.1
+	github.com/redis/go-redis/v9 v9.5.1
+	github.com/samber/lo v1.39.0
+	github.com/sirupsen/logrus v1.9.3
+	github.com/valyala/fasthttp v1.52.0
+	golang.org/x/crypto v0.21.0
+	gopkg.in/guregu/null.v4 v4.0.0
+	gorm.io/gorm v1.25.8
 )
 
 require (
-	github.com/cespare/xxhash/v2 v2.1.1 // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/dgryski/go-rendezvous v0.0.0-20200624174652-8d2f3be8b2d9 // indirect
-	github.com/gin-contrib/sse v0.1.0 // indirect
-	github.com/go-playground/locales v0.13.0 // indirect
-	github.com/go-playground/universal-translator v0.17.0 // indirect
-	github.com/go-playground/validator/v10 v10.4.1 // indirect
-	github.com/go-session/session v3.1.2+incompatible // indirect
-	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
-	github.com/golang/protobuf v1.4.2 // indirect
-	github.com/gopherjs/gopherjs v0.0.0-20190812055157-5d271430af9f // indirect
-	github.com/konsorten/go-windows-terminal-sequences v1.0.1 // indirect
-	github.com/leodido/go-urn v1.2.0 // indirect
-	github.com/mattn/go-isatty v0.0.12 // indirect
-	github.com/modern-go/concurrent v0.0.0-20180306012644-bacd9c7ef1dd // indirect
-	github.com/modern-go/reflect2 v1.0.2 // indirect
-	github.com/mohae/deepcopy v0.0.0-20170929034955-c48cc78d4826 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/richardlehane/mscfb v1.0.4 // indirect
-	github.com/richardlehane/msoleps v1.0.1 // indirect
-	github.com/smartystreets/goconvey v0.0.0-20190731233626-505e41936337 // indirect
-	github.com/ugorji/go/codec v1.1.7 // indirect
-	github.com/xuri/efp v0.0.0-20220407160117-ad0f7a785be8 // indirect
-	github.com/xuri/nfp v0.0.0-20220409054826-5e722a1d9e22 // indirect
-	go.opentelemetry.io/otel v0.7.0 // indirect
-	golang.org/x/exp v0.0.0-20200513190911-00229845015e // indirect
-	golang.org/x/image v0.5.0 // indirect
-	golang.org/x/net v0.7.0 // indirect
-	golang.org/x/sys v0.5.0 // indirect
-	golang.org/x/text v0.7.0 // indirect
-	google.golang.org/grpc v1.30.0 // indirect
-	google.golang.org/protobuf v1.23.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	github.com/andybalholm/brotli v1.1.0 // indirect
+	github.com/cespare/xxhash/v2 v2.2.0 // indirect
+	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
+	github.com/go-playground/locales v0.14.1 // indirect
+	github.com/go-playground/universal-translator v0.18.1 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+	github.com/klauspost/compress v1.17.6 // indirect
+	github.com/leodido/go-urn v1.4.0 // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/mattn/go-runewidth v0.0.15 // indirect
+	github.com/rivo/uniseg v0.2.0 // indirect
+	github.com/valyala/bytebufferpool v1.0.0 // indirect
+	github.com/valyala/tcplisten v1.0.0 // indirect
+	golang.org/x/exp v0.0.0-20220303212507-bbda1eaf7a17 // indirect
+	golang.org/x/net v0.21.0 // indirect
+	golang.org/x/sys v0.18.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	gopkg.in/natefinch/lumberjack.v2 v2.2.1 // indirect
 )
-
-go 1.18
